Generalize ValidatorListTable
