+++ src/pages/staking/staking.tsx
@@ -12,7 +12,6 @@ import {
   Table,
   Tabs,
   Typography,
-  Tooltip,
   Alert,
 } from 'antd';
 import { OrderedListOutlined, ExclamationCircleOutlined } from '@ant-design/icons';
@@ -37,7 +36,6 @@ import {
   getAssetStakingBalancePrice,
   getAssetUnbondingBalancePrice,
   getAssetRewardsBalancePrice,
-  scaledAmount,
   scaledBalance,
   scaledStakingBalance,
   scaledUnbondingBalance,
@@ -47,7 +45,6 @@ import {
 import {
   BroadCastResult,
   RewardTransaction,
-  ValidatorModel,
   StakingTransactionData,
   UnbondingDelegationData,
 } from '../../models/Transaction';
@@ -59,7 +56,7 @@ import {
   getCurrentMinAssetAmount,
   getUIDynamicAmount,
 } from '../../utils/NumberUtils';
-import { middleEllipsis, ellipsis } from '../../utils/utils';
+import { middleEllipsis } from '../../utils/utils';
 import { LEDGER_WALLET_TYPE, detectConditionsError } from '../../service/LedgerService';
 import {
   AnalyticsActions,
@@ -74,19 +71,18 @@ import ModalPopup from '../../components/ModalPopup/ModalPopup';
 import SuccessModalPopup from '../../components/SuccessModalPopup/SuccessModalPopup';
 import ErrorModalPopup from '../../components/ErrorModalPopup/ErrorModalPopup';
 import PasswordFormModal from '../../components/PasswordForm/PasswordFormModal';
-import { UndelegateFormComponent } from '../home/components/UndelegateFormComponent';
-import RedelegateFormComponent from '../home/components/RedelegateFormComponent';
-import ValidatorPowerPercentBar from '../../components/ValidatorPowerPercentBar/ValidatorPowerPercentBar';
+import { UndelegateFormComponent } from './components/UndelegateFormComponent';
+import RedelegateFormComponent from './components/RedelegateFormComponent';
 import RowAmountOption from '../../components/RowAmountOption/RowAmountOption';
+import ValidatorListTable from './components/ValidatorListTable';
+
 import {
   MODERATION_CONFIG_FILE_URL,
   UNBLOCKING_PERIOD_IN_DAYS,
-  CUMULATIVE_SHARE_PERCENTAGE_THRESHOLD,
   FIXED_DEFAULT_FEE,
   SUPPORTED_CURRENCY,
 } from '../../config/StaticConfig';
 import { isValidatorAddressSuspicious, ModerationConfig } from '../../models/ModerationConfig';
-import { ChainIndexingAPI } from '../../service/rpc/ChainIndexingAPI';
 
 const { Header, Content, Footer, Sider } = Layout;
 const { Search } = Input;
@@ -130,7 +126,8 @@ interface UnbondingDelegationTabularData {
   completionTime: string;
 }
 
-const FormDelegationRequest = () => {
+const FormDelegationRequest = props => {
+  const { moderationConfig } = props;
   const [form] = Form.useForm();
   const [formValues, setFormValues] = useState({
     validatorAddress: '',
@@ -155,9 +152,7 @@ const FormDelegationRequest = () => {
   const fetchingDB = useRecoilValue(fetchingDBState);
 
   const [ledgerIsExpertMode, setLedgerIsExpertMode] = useRecoilState(ledgerIsExpertModeState);
-  const [validatorTopList, setValidatorTopList] = useState<ValidatorModel[]>([]);
-  const [displayWarning, setDisplayWarning] = useState(true);
-  const [moderationConfig, setModerationConfig] = useState<ModerationConfig>();
+  // const [moderationConfig, setModerationConfig] = useState<ModerationConfig>();
 
   const analyticsService = new AnalyticsService(currentSession);
 
@@ -166,75 +161,30 @@ const FormDelegationRequest = () => {
 
   const [t] = useTranslation();
 
-  const processValidatorList = async (validatorList: ValidatorModel[] | null) => {
-    if (validatorList) {
-      let willDisplayWarningColumn = false;
-      let displayedWarningColumn = false;
-      const apiClient = ChainIndexingAPI.init(currentSession.wallet.config.indexingUrl);
-
-      return await Promise.all(
-        validatorList.map(async (validator, idx) => {
-          if (
-            new Big(validator.cumulativeSharesIncludePercentage!).gte(
-              CUMULATIVE_SHARE_PERCENTAGE_THRESHOLD,
-            ) &&
-            !displayedWarningColumn
-          ) {
-            displayedWarningColumn = true;
-            willDisplayWarningColumn = true;
-          }
-          const [validatorApyRaw, validatorUptime] = await Promise.all([
-            apiClient.getValidatorsAverageApy([validator.validatorAddress]),
-            apiClient.getValidatorUptimeByAddress(validator.validatorAddress),
-          ]);
-
-          const validatorModel = {
-            ...validator,
-            apy: validatorApyRaw ? new Big(validatorApyRaw).toString() : '0',
-            uptime: validatorUptime ? new Big(validatorUptime).toString() : '0',
-            key: `${idx}`,
-            displayWarningColumn: willDisplayWarningColumn,
-          };
-
-          willDisplayWarningColumn = false;
-
-          return validatorModel;
-        }),
-      );
-    }
-    return [];
-  };
-
   useEffect(() => {
-    const syncValidatorsData = async () => {
-      const validatorList = await processValidatorList(currentValidatorList);
-      setValidatorTopList(validatorList);
-    };
-
     const syncAssetData = async () => {
       const currentWalletAsset = await walletService.retrieveDefaultWalletAsset(currentSession);
       setWalletAsset(currentWalletAsset);
     };
 
-    const moderationConfigHandler = async () => {
-      try {
-        const fetchModerationConfigData = await fetch(MODERATION_CONFIG_FILE_URL);
-        const moderationConfigData = await fetchModerationConfigData.json();
-        setModerationConfig(moderationConfigData);
-      } catch (error) {
-        // eslint-disable-next-line no-console
-        console.error('Error occurred while fetching moderation config file', error);
-      }
-    };
+    // const moderationConfigHandler = async () => {
+    //   try {
+    //     const fetchModerationConfigData = await fetch(MODERATION_CONFIG_FILE_URL);
+    //     const moderationConfigData = await fetchModerationConfigData.json();
+    //     setModerationConfig(moderationConfigData);
+    //   } catch (error) {
+    //     // eslint-disable-next-line no-console
+    //     console.error('Error occurred while fetching moderation config file', error);
+    //   }
+    // };
 
     if (!didMountRef.current) {
       syncAssetData();
       didMountRef.current = true;
     }
 
-    syncValidatorsData();
-    moderationConfigHandler();
-  }, [fetchingDB, walletAsset, currentValidatorList]);
+    // moderationConfigHandler();
+  }, [fetchingDB, walletAsset]);
 
   const showConfirmationModal = () => {
     setInputPasswordVisible(false);
@@ -358,126 +308,6 @@ const FormDelegationRequest = () => {
     }`,
   );
 
-  const validatorColumns = [
-    {
-      title: t('staking.validatorList.table.validatorName'),
-      dataIndex: 'validatorName',
-      key: 'validatorName',
-      render: (validatorName, record: ValidatorModel) => (
-        <a
-          data-original={record.validatorAddress}
-          target="_blank"
-          rel="noreferrer"
-          href={`${renderExplorerUrl(currentSession.wallet.config, 'validator')}/${
-            record.validatorAddress
-          }`}
-        >
-          {ellipsis(validatorName, 24)}{' '}
-          {isValidatorAddressSuspicious(record.validatorAddress, moderationConfig) && (
-            <Tooltip title={t('staking.model1.warning')}>
-              <span>
-                <ExclamationCircleOutlined style={{ color: 'red' }} />
-              </span>
-            </Tooltip>
-          )}
-        </a>
-      ),
-    },
-    {
-      title: t('staking.validatorList.table.validatorAddress'),
-      dataIndex: 'validatorAddress',
-      key: 'validatorAddress',
-      render: validatorAddress => (
-        <a
-          data-original={validatorAddress}
-          title={validatorAddress}
-          target="_blank"
-          rel="noreferrer"
-          href={`${renderExplorerUrl(
-            currentSession.wallet.config,
-            'validator',
-          )}/${validatorAddress}`}
-        >
-          {middleEllipsis(validatorAddress, 6)}
-        </a>
-      ),
-    },
-    {
-      title: t('staking.validatorList.table.currentTokens'),
-      dataIndex: 'currentTokens',
-      key: 'currentTokens',
-      sorter: (a, b) => new Big(a.currentTokens).cmp(new Big(b.currentTokens)),
-      defaultSortOrder: 'descend' as any,
-      render: currentTokens => {
-        return (
-          <span>
-            {numeral(scaledAmount(currentTokens, 8)).format('0,0')}{' '}
-            {currentSession.wallet.config.network.coin.croDenom.toUpperCase()}
-          </span>
-        );
-      },
-    },
-    {
-      title: t('staking.validatorList.table.cumulativeShares'),
-      // dataIndex: 'cumulativeShares',
-      key: 'cumulativeShares',
-      // sorter: (a, b) => new Big(a.cumulativeShares).cmp(new Big(b.cumulativeShares)),
-      defaultSortOrder: 'descend' as any,
-      render: record => {
-        return (
-          <>
-            {/* <span>{record.cumulativeShares} %</span> */}
-            <ValidatorPowerPercentBar
-              percentExcludeCurrent={record.cumulativeSharesExcludePercentage} // light blue
-              percentIncludeCurrent={record.cumulativeSharesIncludePercentage} // primary blue
-            />
-          </>
-        );
-      },
-    },
-    {
-      title: t('staking.validatorList.table.currentCommissionRate'),
-      dataIndex: 'currentCommissionRate',
-      key: 'currentCommissionRate',
-      sorter: (a, b) => new Big(a.currentCommissionRate).cmp(new Big(b.currentCommissionRate)),
-      render: currentCommissionRate => (
-        <span>{new Big(currentCommissionRate).times(100).toFixed(2)}%</span>
-      ),
-    },
-    {
-      title: t('staking.validatorList.table.validatorApy'),
-      key: 'apy',
-      sorter: (a, b) => new Big(a.apy).cmp(new Big(b.apy)),
-      render: record => {
-        return <span>{new Big(record.apy).times(100).toFixed(2)}%</span>;
-      },
-    },
-    {
-      title: t('staking.validatorList.table.validatorUptime'),
-      key: 'uptime',
-      sorter: (a, b) => new Big(a.uptime).cmp(new Big(b.uptime)),
-      render: record => {
-        return <span>{new Big(record.uptime).times(100).toFixed(2)}%</span>;
-      },
-    },
-    {
-      title: t('general.action'),
-      key: 'action',
-      render: record => (
-        <a
-          onClick={() => {
-            setIsValidatorListVisible(false);
-            form.setFieldsValue({
-              validatorAddress: record.validatorAddress,
-            });
-          }}
-        >
-          {t('general.select')}
-        </a>
-      ),
-    },
-  ];
-
   function onShowMemoChange() {
     setShowMemo(!showMemo);
   }
@@ -511,44 +341,13 @@ const FormDelegationRequest = () => {
           <div className="title">{t('staking.validatorList.table.title')}</div>
           <div className="description">{t('staking.validatorList.table.description')}</div>
           <div className="item">
-            <Table
-              locale={{
-                triggerDesc: t('general.table.triggerDesc'),
-                triggerAsc: t('general.table.triggerAsc'),
-                cancelSort: t('general.table.cancelSort'),
-              }}
-              dataSource={validatorTopList}
-              columns={validatorColumns}
-              pagination={{ showSizeChanger: false }}
-              onChange={(pagination, filters, sorter: any) => {
-                if (
-                  (sorter.order === 'descend' && sorter.field === 'currentTokens') ||
-                  sorter.order === undefined
-                ) {
-                  setDisplayWarning(true);
-                } else {
-                  setDisplayWarning(false);
-                }
-              }}
-              expandable={{
-                rowExpandable: record => record.displayWarningColumn! && displayWarning,
-                expandedRowRender: record =>
-                  record.displayWarningColumn &&
-                  displayWarning && (
-                    <div className="cumulative-stake33">
-                      {t('staking.validatorList.table.warningCumulative')}
-                    </div>
-                  ),
-                expandIconColumnIndex: -1,
-              }}
-              rowClassName={record => {
-                const greyBackground =
-                  new Big(record.cumulativeSharesIncludePercentage!).lte(
-                    CUMULATIVE_SHARE_PERCENTAGE_THRESHOLD,
-                  ) || record.displayWarningColumn;
-                return greyBackground ? 'grey-background' : '';
-              }}
-              defaultExpandAllRows
+            <ValidatorListTable
+              currentSession={currentSession}
+              // validatorTopList={validatorTopList}
+              currentValidatorList={currentValidatorList}
+              moderationConfig={moderationConfig}
+              setIsValidatorListVisible={setIsValidatorListVisible}
+              form={form}
             />
           </div>
         </ModalPopup>
@@ -780,7 +579,8 @@ const FormDelegationRequest = () => {
   );
 };
 
-const FormDelegationOperations = () => {
+const FormDelegationOperations = props => {
+  const { moderationConfig } = props;
   // Undelegate action related states changes
   const [form] = Form.useForm();
   const [userAsset, setUserAsset] = useRecoilState(walletAssetState);
@@ -1103,6 +903,7 @@ const FormDelegationOperations = () => {
             <RedelegateFormComponent
               currentSession={currentSession}
               redelegateFormValues={redelegateFormValues}
+              moderationConfig={moderationConfig}
               walletAsset={userAsset}
               form={form}
             />
@@ -1515,14 +1316,18 @@ const FormWithdrawStakingReward = () => {
 const StakingPage = () => {
   const currentSession = useRecoilValue(sessionState);
   const userAsset = useRecoilValue(walletAssetState);
+  const currentValidatorList = useRecoilValue(validatorListState);
   const fetchingDB = useRecoilValue(fetchingDBState);
   const allMarketData = useRecoilValue(allMarketState);
+  // const [validatorTopList, setValidatorTopList] = useState<ValidatorModel[]>([]);
   const [marketData, setMarketData] = useState<AssetMarketPrice>();
+
   const [isUnbondingDelegationModalVisible, setIsUnbondingDelegationModalVisible] = useState(false);
   const [isUnbondingVisible, setIsUnbondingVisible] = useState(false);
   const [unbondingDelegations, setUnbondingDelegations] = useState<
     UnbondingDelegationTabularData[]
   >([]);
+  const [moderationConfig, setModerationConfig] = useState<ModerationConfig>();
   const analyticsService = new AnalyticsService(currentSession);
   const didMountRef = useRef(false);
 
@@ -1606,7 +1411,51 @@ const StakingPage = () => {
     // .filter(dlg => Number(dlg.stakedAmount) > 0);
   };
 
+  // const processValidatorList = async (validatorList: ValidatorModel[] | null) => {
+  //   if (validatorList) {
+  //     let willDisplayWarningColumn = false;
+  //     let displayedWarningColumn = false;
+  //     const apiClient = ChainIndexingAPI.init(currentSession.wallet.config.indexingUrl);
+
+  //     return await Promise.all(
+  //       validatorList.map(async (validator, idx) => {
+  //         if (
+  //           new Big(validator.cumulativeSharesIncludePercentage!).gte(
+  //             CUMULATIVE_SHARE_PERCENTAGE_THRESHOLD,
+  //           ) &&
+  //           !displayedWarningColumn
+  //         ) {
+  //           displayedWarningColumn = true;
+  //           willDisplayWarningColumn = true;
+  //         }
+  //         const [validatorApyRaw, validatorUptime] = await Promise.all([
+  //           apiClient.getValidatorsAverageApy([validator.validatorAddress]),
+  //           apiClient.getValidatorUptimeByAddress(validator.validatorAddress),
+  //         ]);
+
+  //         const validatorModel = {
+  //           ...validator,
+  //           apy: validatorApyRaw ? new Big(validatorApyRaw).toString() : '0',
+  //           uptime: validatorUptime ? new Big(validatorUptime).toString() : '0',
+  //           key: `${idx}`,
+  //           displayWarningColumn: willDisplayWarningColumn,
+  //         };
+
+  //         willDisplayWarningColumn = false;
+
+  //         return validatorModel;
+  //       }),
+  //     );
+  //   }
+  //   return [];
+  // };
+
   useEffect(() => {
+    // const syncValidatorsData = async () => {
+    //   const validatorList = await processValidatorList(currentValidatorList);
+    //   setValidatorTopList(validatorList);
+    // };
+
     const syncUnbondingDelegationsData = async () => {
       const allUnbonding = await walletService.retrieveAllUnbondingDelegations(
         currentSession.wallet.identifier,
@@ -1618,15 +1467,28 @@ const StakingPage = () => {
       setIsUnbondingVisible(unbondingDelegationTabularData.length > 0);
     };
 
+    const moderationConfigHandler = async () => {
+      try {
+        const fetchModerationConfigData = await fetch(MODERATION_CONFIG_FILE_URL);
+        const moderationConfigData = await fetchModerationConfigData.json();
+        setModerationConfig(moderationConfigData);
+      } catch (error) {
+        // eslint-disable-next-line no-console
+        console.error('Error occurred while fetching moderation config file', error);
+      }
+    };
+
     syncUnbondingDelegationsData();
 
     setMarketData(allMarketData[`${userAsset?.mainnetSymbol}-${currentSession.currency}`]);
 
     if (!didMountRef.current) {
       didMountRef.current = true;
+      // syncValidatorsData();
+      moderationConfigHandler();
       analyticsService.logPage('Staking');
     }
-  }, [fetchingDB]);
+  }, [fetchingDB, currentValidatorList]);
 
   return (
     <Layout className="site-layout">
@@ -1712,7 +1574,7 @@ const StakingPage = () => {
             <div className="site-layout-background stake-content">
               <div className="container">
                 <div className="description">{t('staking.description2')}</div>
-                <FormDelegationRequest />
+                <FormDelegationRequest moderationConfig={moderationConfig} />
               </div>
             </div>
           </TabPane>
@@ -1763,7 +1625,7 @@ const StakingPage = () => {
                     <></>
                   )}
                 </Layout>
-                <FormDelegationOperations />
+                <FormDelegationOperations moderationConfig={moderationConfig} />
               </div>
             </div>
           </TabPane>
